[CHORE] 모델 분리 (280)
