+++ Maddori.Apple/Maddori.Apple/Screen/Home/Home/Model/Option.swift
@@ -1,0 +1,13 @@
+//
+//  Option.swift
+//  Maddori.Apple
+//
+//  Created by 이성호 on 2023/02/06.
+//
+
+import Foundation
+
+struct Option {
+    let title: String
+    let handler: () -> Void
+}
