+++ Maddori.Apple/Maddori.Apple/Screen/Home/Home/Model/Section.swift
@@ -1,0 +1,13 @@
+//
+//  Section.swift
+//  Maddori.Apple
+//
+//  Created by 이성호 on 2023/02/06.
+//
+
+import Foundation
+
+struct Section {
+    let title: String
+    let options: [Option]
+}
